requirements.txt
gpiozero~=1.6.2
paho-mqtt~=1.5.1
PyYAML~=5.4.1
tqdm~=4.62.3
